requirements.txt
selenium
python-dotenv
python-telegram-bot
schedule
holidays
requests
reportlab
pandas
matplotlib
seaborn
fpdf
xlsxwriter
